chore(build): organize vite/ into configs/ and scripts/ subdirs
  vite/configs/worker-configs.{esm,umd}.js
  vite/scripts/build-bundle-{esm,umd}.js
