+++ package.json
@@ -48,8 +48,8 @@
     "build": "npm run build:app && npm run build:bundle:esm && npm run build:bundle:umd && npm run build:definitions",
     "build:app": "cross-env NODE_OPTIONS=--max_old_space_size=8192 vite build --config vite.config.app.js",
     "build:app:serve": "vite preview --outDir build --port 3050",
-    "build:bundle:esm": "rimraf ./dist/esm && cross-env NODE_OPTIONS=--max_old_space_size=4096 node vite/build-bundle-esm.js && rimraf ./dist/esm/swagger-editor.css*",
-    "build:bundle:umd": "rimraf ./dist/umd ./dist/swagger-editor.css && cross-env NODE_OPTIONS=--max_old_space_size=4096 node vite/build-bundle-umd.js && copyfiles -u 2 ./dist/umd/swagger-editor.css ./dist && rimraf ./dist/umd/swagger-editor.css",
+    "build:bundle:esm": "rimraf ./dist/esm && cross-env NODE_OPTIONS=--max_old_space_size=4096 node vite/scripts/build-bundle-esm.js && rimraf ./dist/esm/swagger-editor.css*",
+    "build:bundle:umd": "rimraf ./dist/umd ./dist/swagger-editor.css && cross-env NODE_OPTIONS=--max_old_space_size=4096 node vite/scripts/build-bundle-umd.js && copyfiles -u 2 ./dist/umd/swagger-editor.css ./dist && rimraf ./dist/umd/swagger-editor.css",
     "build:definitions": "tsc -p tsconfig.json",
     "test": "cross-env NODE_ENV=test BABEL_ENV=test PUBLIC_URL= node scripts/test.js",
     "pw:install": "playwright install --with-deps chromium",
