+++ vite/configs/worker-configs.esm.js
@@ -2,8 +2,8 @@ import { defineConfig } from 'vite';
 import { resolve, dirname } from 'path';
 import { fileURLToPath } from 'url';
 
-import { logger, sharedOnwarn } from './shared.js';
-import { inlineAllWasms } from './plugins/inline-all-wasms.js';
+import { logger, sharedOnwarn } from '../shared.js';
+import { inlineAllWasms } from '../plugins/inline-all-wasms.js';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
@@ -18,7 +18,7 @@ export const apidomWorkerConfig = defineConfig({
 
   build: {
     lib: {
-      entry: resolve(__dirname, '../src/plugins/editor-monaco-language-apidom/language/apidom.worker.js'),
+      entry: resolve(__dirname, '../../src/plugins/editor-monaco-language-apidom/language/apidom.worker.js'),
       formats: ['es'],
       fileName: () => 'apidom.worker.js',
     },
@@ -40,7 +40,7 @@ export const editorWorkerConfig = defineConfig({
 
   build: {
     lib: {
-      entry: resolve(__dirname, '../node_modules/monaco-editor/esm/vs/editor/editor.worker.js'),
+      entry: resolve(__dirname, '../../node_modules/monaco-editor/esm/vs/editor/editor.worker.js'),
       formats: ['es'],
       fileName: () => 'editor.worker.js',
     },
