+++ vite/configs/worker-configs.umd.js
@@ -2,8 +2,8 @@ import { defineConfig } from 'vite';
 import { resolve, dirname } from 'path';
 import { fileURLToPath } from 'url';
 
-import { logger, sharedOnwarn } from './shared.js';
-import { inlineAllWasms } from './plugins/inline-all-wasms.js';
+import { logger, sharedOnwarn } from '../shared.js';
+import { inlineAllWasms } from '../plugins/inline-all-wasms.js';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
@@ -15,7 +15,7 @@ export const apidomWorkerConfig = defineConfig({
 
   build: {
     lib: {
-      entry: resolve(__dirname, '../src/plugins/editor-monaco-language-apidom/language/apidom.worker.js'),
+      entry: resolve(__dirname, '../../src/plugins/editor-monaco-language-apidom/language/apidom.worker.js'),
       formats: ['iife'],
       fileName: () => 'apidom.worker.js',
       name: 'apidomWorker',
@@ -38,7 +38,7 @@ export const editorWorkerConfig = defineConfig({
 
   build: {
     lib: {
-      entry: resolve(__dirname, '../node_modules/monaco-editor/esm/vs/editor/editor.worker.js'),
+      entry: resolve(__dirname, '../../node_modules/monaco-editor/esm/vs/editor/editor.worker.js'),
       formats: ['iife'],
       fileName: () => 'editor.worker.js',
       name: 'editorWorker',
